Module epic_wallet::cmd::wallet_args

Enums

ParseError

Simple error definition, just so we can return errors from all commands and let the caller figure out what to do

Functions

inst_wallet

parse_account_args

parse_cancel_args

parse_check_args

parse_finalize_args

parse_global_args

parse_info_args

parse_init_args

parse_issue_invoice_args

parse_listen_args

parse_owner_api_args

parse_post_args

parse_process_invoice_args

parse_receive_args

parse_recover_args

parse_repost_args

parse_send_args

parse_txs_args

prompt_password

wallet_command